requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Admin
Flask-Cors
Flask-Bcrypt
python-dotenv
psycopg2
